requirements.txt
fastapi
uvicorn[standard]
torch
numpy
pydantic
wandb
streamlit
streamlit-drawable-canvas
Pillow
requests
psycopg2-binary
sqlalchemy
python-dotenv








